requirements.txt
asgiref==3.2.7
astroid==2.3.3
autopep8==1.5
colorama==0.4.3
Django==3.0.4
django-cors-headers==3.2.1
django-debug-toolbar==2.2
django-filter==2.2.0
django-rest-framework==0.1.0
django-webpack-loader==0.7.0
djangorestframework==3.11.0
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
pycodestyle==2.5.0
PyJWT==1.7.1
pylint==2.4.4
pytz==2019.3
six==1.14.0
sqlparse==0.3.1
typed-ast==1.4.1
wrapt==1.11.2
